requirements.txt
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.9.2
click==8.1.3
flask==2.2.2
flask-cors==3.0.10
flask-login==0.6.2
flask-migrate==4.0.2
flask-sqlalchemy==3.0.2
flask-wtf==1.1.1
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
jinja2==3.1.2
mako==1.2.4
markupsafe==2.1.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
six==1.16.0
sqlalchemy==1.4.46
werkzeug==2.2.2
wtforms==3.0.1
psycopg2==2.9.9
boto3==1.33.13
google-auth==2.3.3
google-api-python-client==2.34.0